ASC-760 Test attachment of junit xml and text log
Two artifacts should be attached to the swagger
client so they can be uploaded to qtest for use
in failure analysis: the junit xml file, and
a text representation of the failure log.

This commit adds a test to ensure that:
there is a list of two artifacts,
and that those files have the correct mime types.