ASC-761 add a failure log summary
Prior to this change a very long log output string could be
put in the qtest failure output field. This can throw off the
page layout.

This update truncates the log if it's over some character limit
and refers the user to the attached text file containing the
log output.