ASC-918 Add Support for Test Case Steps
Test cases marked with 'test_step' property marked with 'true' in the JUnitXML
will be treated as test steps for a combined test case with test case name
derived from the 'classname' JUnitXML attribute.

This feature required a ton of changes. As I worked through the changes I found
a bunch of small formatting issues that I fixed along the way. The diff is
super gross. Sorry | not sorry?